Rename deleteModal.tsx
