+++ workspaces/frontend/src/shared/components/DeleteModal.tsx
